# docs/v1/swagger.yaml
basePath: /v1
definitions:
...
  protocol.DataType:
    enum:
    - object
    - number
    - integer
    - string
    - array
    - "null"
    - boolean
    type: string
    x-enum-varnames:
    - ObjectT
    - Number
    - Integer
    - String
    - Array
    - "Null"
    - Boolean
  protocol.InputSchema:
    properties:
      properties:
        additionalProperties:
          $ref: '#/definitions/protocol.Property'
        type: object
      required:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/protocol.InputSchemaType'
    type: object
  protocol.InputSchemaType:
    enum:
    - object
    type: string
    x-enum-varnames:
    - Object
  protocol.OutputSchema:
    properties:
      properties:
        additionalProperties:
          $ref: '#/definitions/protocol.Property'
        type: object
      required:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/protocol.InputSchemaType'
    type: object
  protocol.Property:
    properties:
      description:
        description: Description is the description of the schema.
        type: string
      enum:
        items:
          type: string
        type: array
      items:
        allOf:
        - $ref: '#/definitions/protocol.Property'
        description: Items specifies which data type an array contains, if the schema
          type is Array.
      properties:
        additionalProperties:
          $ref: '#/definitions/protocol.Property'
        description: Properties describes the properties of an object, if the schema
          type is Object.
        type: object
      required:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/protocol.DataType'
    type: object
  protocol.Tool:
    properties:
      annotations:
        allOf:
        - $ref: '#/definitions/protocol.ToolAnnotations'
        description: Annotations provides additional hints about the tool's behavior
      description:
        description: Description is a human-readable description of the tool
        type: string
      inputSchema:
        allOf:
        - $ref: '#/definitions/protocol.InputSchema'
        description: InputSchema defines the expected parameters for the tool using
          JSON Schema
      name:
        description: Name is the unique identifier of the tool
        type: string
      outputSchema:
        allOf:
        - $ref: '#/definitions/protocol.OutputSchema'
        description: OutputSchema defines expected output structure for the tool using
          Optional JSON Schema
    type: object
  protocol.ToolAnnotations:
    properties:
      destructiveHint:
        description: |-
          DestructiveHint indicates the tool may perform destructive updates
          (only meaningful when ReadOnlyHint is false)
        type: boolean
      idempotentHint:
        description: |-
          IdempotentHint indicates calling the tool repeatedly with the same arguments
          has no additional effect (only meaningful when ReadOnlyHint is false)
        type: boolean
      openWorldHint:
        description: OpenWorldHint indicates the tool may interact with an "open world"
          of external entities
        type: boolean
      readOnlyHint:
        description: ReadOnlyHint indicates the tool does not modify its environment
        type: boolean
      title:
        description: Title is a human-readable title for the tool, useful for UI display
        type: string
    type: object
...
  request.Avatar:
    properties:
      key:
        description: 前端透传给后端用于保存avatar，例如：custom-upload/avatar/abc/def.png
        type: string
      path:
        description: 前端请求地址，例如：/v1/static/avatar/abc/def.png (请求非必填)
        type: string
    type: object
...
  request.CustomToolCreate:
    properties:
      apiAuth:
        allOf:
        - $ref: '#/definitions/util.ApiAuthWebRequest'
        description: api身份认证
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      description:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
      privacyPolicy:
        description: 隐私政策
        type: string
      schema:
        description: schema
        type: string
    required:
    - apiAuth
    - description
    - name
    - schema
    type: object
  request.CustomToolIDReq:
    properties:
      customToolId:
        description: 自定义工具id
        type: string
    required:
    - customToolId
    type: object
  request.CustomToolSchemaReq:
    properties:
      schema:
        description: schema
        type: string
    required:
    - schema
    type: object
  request.CustomToolUpdateReq:
    properties:
      apiAuth:
        allOf:
        - $ref: '#/definitions/util.ApiAuthWebRequest'
        description: api身份认证
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      customToolId:
        description: 自定义工具ID
        type: string
      description:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
      privacyPolicy:
        description: 隐私政策
        type: string
      schema:
        description: schema
        type: string
    required:
    - apiAuth
    - customToolId
    - description
    - name
    - schema
    type: object
...
  request.ToolSquareAPIKeyReq:
    properties:
      apiKey:
        description: apiKey
        type: string
      toolSquareId:
        description: 广场toolId
        type: string
    required:
    - toolSquareId
    type: object
...
  response.CustomToolActionInfo:
    properties:
      desc:
        type: string
      method:
        type: string
      name:
        type: string
      path:
        type: string
    type: object
  response.CustomToolDetail:
    properties:
      apiAuth:
        allOf:
        - $ref: '#/definitions/util.ApiAuthWebRequest'
        description: apiAuth
      apiList:
        description: action列表
        items:
          $ref: '#/definitions/response.CustomToolActionInfo'
        type: array
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      customToolId:
        description: 自定义工具id
        type: string
      description:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
      privacyPolicy:
        description: 隐私政策
        type: string
      schema:
        description: schema
        type: string
    type: object
  response.CustomToolInfo:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      customToolId:
        description: 自定义工具id
        type: string
      description:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
    type: object
...
  response.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.ToolActionDetail:
    properties:
      action:
        allOf:
        - $ref: '#/definitions/protocol.Tool'
        description: action列表
      apiKey:
        description: apiKey
        type: string
      needApiKeyInput:
        description: 是否需要apiKey输入
        type: boolean
    type: object
  response.ToolActionList:
    properties:
      actions:
        description: action列表
        items:
          $ref: '#/definitions/protocol.Tool'
        type: array
    type: object
...
  response.ToolSelect:
    properties:
      apiKey:
        description: apiKey
        type: string
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 工具描述
        type: string
      needApiKeyInput:
        description: 是否需要apiKey输入
        type: boolean
      toolId:
        description: 工具id
        type: string
      toolName:
        description: 工具名称
        type: string
      toolType:
        description: 工具类型
        enum:
        - custom
        - builtin
        type: string
      uniqueId:
        description: unique id
        type: string
    required:
    - toolType
    type: object
...
  response.ToolSquareDetail:
    properties:
      actionSum:
        description: action总数
        type: integer
      apiKey:
        description: apiKey
        type: string
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      detail:
        description: 详细描述
        type: string
      name:
        description: 名称
        type: string
      needApiKeyInput:
        description: 是否需要apiKey输入
        type: boolean
      schema:
        type: string
      tags:
        description: 标签
        items:
          type: string
        type: array
      toolSquareId:
        description: 广场mcpId(非空表示来源于广场)
        type: string
      tools:
        description: action列表
        items:
          $ref: '#/definitions/protocol.Tool'
        type: array
    type: object
  response.ToolSquareInfo:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
      tags:
        description: 标签
        items:
          type: string
        type: array
      toolSquareId:
        description: 广场mcpId(非空表示来源于广场)
        type: string
    type: object
...
  util.ApiAuthWebRequest:
    properties:
      apiKeyHeader:
        description: HTTP头部名称
        type: string
      apiKeyHeaderPrefix:
        description: 鉴权头部前缀
        enum:
        - basic
        - bearer
        - custom
        type: string
      apiKeyQueryParam:
        description: 查询参数名称
        type: string
      apiKeyValue:
        description: apiKey
        type: string
      authType:
        description: 鉴权类型 None 或 请求头 或 查询参数
        enum:
        - none
        - api_key_query
        - api_key_header
        type: string
    required:
    - authType
    type: object
info:
  contact: {}
  description: |-
    ## HTTP Header
    | Header        | 说明      |
    | ------------- | --------- |
    | Authorization | JWT token |
    | X-Language    | 语言Code  |
    | X-Org-Id      | 组织ID    |
    | X-Client-Id   | 客户端标识|

    ## HTTP Status
    | HTTP Status             | 说明                   |
    | ----------------------- | ---------------------- |
    | 200, StatusOK           | 请求返回成功           |
    | 400, StatusBadRequest   | 请求返回失败，用于业务 |
    | 401, StatusUnauthorized | JWT认证失败            |
    | 403, StatusForbidden    | 没有权限               |

    ## 权限-菜单对应表
    | 一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |
    |-------------|-------|------|------|------|------|
    | guest       |       |      | 【访客】 |      |      |
    | common      |       |      | 【通用】 |      |      |
    | permission  |       |      | 权限管理 |      |      |
    | permission  | user  |      | 权限管理 | 用户管理 |      |
    | permission  | org   |      | 权限管理 | 组织管理 |      |
    | permission  | role  |      | 权限管理 | 角色管理 |      |

    ## `/v1/user/permission`返回用例
    ```json
    {
      "code": 0,
      "data": {
        "orgPermission": {
          "org": {"id": "test-org-id", "name": "test-org-name"},
          "permissions": [
            {"perm": "permission"},
            {"perm": "permission.user"},
            {"perm": "permission.org"},
            {"perm": "permission.role"}
          ]
        }
      },
      "msg": "操作成功"
    }
    ```
...
paths:
...
  /tool/action/detail:
    get:
      consumes:
      - application/json
      description: 获取工具详情
      parameters:
      - description: action名称
        in: query
        name: actionName
        required: true
        type: string
      - description: 工具id
        in: query
        name: toolId
        required: true
        type: string
      - description: 工具类型
        enum:
        - builtin
        - custom
        in: query
        name: toolType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ToolActionDetail'
              type: object
      summary: 获取工具详情
      tags:
      - tool
  /tool/action/list:
    get:
      consumes:
      - application/json
      description: 获取工具列表
      parameters:
      - description: 工具id
        in: query
        name: toolId
        required: true
        type: string
      - description: 工具类型
        enum:
        - builtin
        - custom
        in: query
        name: toolType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ToolActionList'
              type: object
      summary: 获取工具列表
      tags:
      - tool
  /tool/builtin:
    post:
      consumes:
      - application/json
      description: 修改内置工具
      parameters:
      - description: 内置工具信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ToolSquareAPIKeyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改内置工具
      tags:
      - tool
  /tool/custom:
    delete:
      consumes:
      - application/json
      description: 删除自定义工具
      parameters:
      - description: 自定义工具ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CustomToolIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除自定义工具
      tags:
      - tool
    get:
      consumes:
      - application/json
      description: 获取自定义工具详情
      parameters:
      - description: customToolId
        in: query
        name: customToolId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CustomToolDetail'
              type: object
      summary: 获取自定义工具详情
      tags:
      - tool
    post:
      consumes:
      - application/json
      description: 创建自定义工具
      parameters:
      - description: 自定义工具信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CustomToolCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建自定义工具
      tags:
      - tool
    put:
      consumes:
      - application/json
      description: 修改自定义工具
      parameters:
      - description: 自定义工具信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CustomToolUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改自定义工具
      tags:
      - tool
  /tool/custom/list:
    get:
      consumes:
      - application/json
      description: 获取自定义工具列表
      parameters:
      - description: CustomTool名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.CustomToolInfo'
                        type: array
                    type: object
              type: object
      summary: 获取自定义工具列表
      tags:
      - tool
  /tool/custom/schema:
    post:
      consumes:
      - application/json
      description: 解析自定义工具的Schema转换为API相关数据
      parameters:
      - description: Schema格式数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CustomToolSchemaReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.CustomToolActionInfo'
                        type: array
                    type: object
              type: object
      summary: 获取可用API列表（根据Schema）
      tags:
      - tool
...
  /tool/select:
    get:
      consumes:
      - application/json
      description: 获取工具列表（用于下拉选择）
      parameters:
      - description: 工具名
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.ToolSelect'
                        type: array
                    type: object
              type: object
      summary: 获取工具列表（用于下拉选择）
      tags:
      - tool
  /tool/square:
    get:
      consumes:
      - application/json
      description: 获取内置工具详情
      parameters:
      - description: toolSquareId
        in: query
        name: toolSquareId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ToolSquareDetail'
              type: object
      summary: 获取内置工具详情
      tags:
      - tool.square
  /tool/square/list:
    get:
      consumes:
      - application/json
      description: 获取内置工具列表
      parameters:
      - description: tool名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.ToolSquareInfo'
                        type: array
                    type: object
              type: object
      summary: 获取内置工具列表
      tags:
      - tool.square
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"